+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### Fixed
 
-- Mimic the way attributeChangedCallback is dispatched on connectedCallback.
+- Makes propertyChangedCallback mimic the way attributeChangedCallback is dispatched on connectedCallback.
 
 
 ## 0.1.9 - 2018-07-18